Extract button of modules to components
