+++ frontend/components/UpdateSourceModulesButton/index.ts
@@ -1,0 +1,1 @@
+export { UpdateSourceModulesButton } from './UpdateSourceModulesButton'
